Simplify helper files
